pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "py-occam"
version = "0.1.0"
authors = [
  { name="Your Name", email="your.email@example.com" },
]
description = "Python implementation of OCCAM variable-based reconstructability analysis"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.23.0", 
    "scipy>=1.9.0",
    "pyyaml>=6.0",
]

[project.urls]
"Homepage" = "https://github.com/percyd/occam"
"Bug Tracker" = "https://github.com/percyd/occam/issues"